remove app, menu, joined into one file for easier calling of functions
